refactor: DemoDatabaseInitializer profile mockData 추가
